+++ src/Definitions/Polizas13Definition.php
@@ -8,11 +8,11 @@ interface Polizas13Definition
 {
     public const NAMESPACE = 'http://www.sat.gob.mx/esquemas/ContabilidadE/1_3/PolizasPeriodo';
 
-    public const XSD_LOCATION = 'http://www.sat.gob.mx/esquemas/ContabilidadE/1_3/PolizasPeriodo'
-        . '/PolizasPeriodo_1_3.xsd';
+    public const XSD_LOCATION = 'http://www.sat.gob.mx/esquemas/ContabilidadE'
+        . '/1_3/PolizasPeriodo/PolizasPeriodo_1_3.xsd';
 
-    public const XSLT_LOCATION = 'http://www.sat.gob.mx/esquemas/ContabilidadE/1_1/PolizasPeriodo'
-        . '/PolizasPeriodo_1_1.xslt';
+    public const XSLT_LOCATION = 'http://www.sat.gob.mx/esquemas/ContabilidadE'
+        . '/1_3/PolizasPeriodo/PolizasPeriodo_1_2.xslt';
 
     public const ELEMENT_NAME = 'PLZ:Polizas';
 }